telegram_bot/requirements.txt
# telegram-bot/requirements.txt
python-telegram-bot==20.7